# Conflicts:
#	CMakeLists.txt
#	Common/CMakeLists.txt
#	Common/ac/characterinfo.cpp
#	Common/ac/characterinfo.h
#	Common/ac/dialogtopic.cpp
#	Common/ac/game_version.h
#	Common/ac/gamesetupstruct.cpp
#	Common/core/def_version.h
#	Common/game/data_helpers.h
#	Common/game/room_file.cpp
#	Common/game/room_file_base.cpp
#	Common/game/room_version.h
#	Common/game/roomstruct.cpp
#	Common/game/roomstruct.h
#	Common/gui/guibutton.cpp
#	Common/gui/guilabel.cpp
#	Editor/AGS.Editor/AGSEditor.cs
#	Editor/AGS.Editor/AGSEditor.csproj
#	Editor/AGS.Editor/Components/TranslationsComponent.cs
#	Editor/AGS.Editor/Panes/Room/RoomEditFilters/CharactersEditorFilter.cs
#	Editor/AGS.Editor/Panes/Room/RoomEditFilters/ObjectsEditorFilter.cs
#	Editor/AGS.Editor/Tasks.cs
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Native/agsnative.cpp
#	Editor/AGS.Types/AGS.Types.csproj
#	Editor/AGS.Types/Attributes/SerializeAsHex.cs
#	Editor/AGS.Types/Character.cs
#	Editor/AGS.Types/GUIButton.cs
#	Editor/AGS.Types/GUISlider.cs
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/AGS.Types/PropertyGridExtras/ColorUIEditor.cs
#	Editor/AGS.Types/RoomObject.cs
#	Editor/AGS.Types/SerializeUtils.cs
#	Engine/ac/character.cpp
#	Engine/ac/characterextras.cpp
#	Engine/ac/characterextras.h
#	Engine/ac/room.cpp
#	Engine/ac/roomobject.h
#	Engine/game/savegame_components.cpp
#	Engine/main/game_run.cpp
#	Solutions/Common.Lib/Common.Lib.vcxproj
#	Solutions/Common.Lib/Common.Lib.vcxproj.filters
#	version.json

+++ Engine/ac/character.cpp
@@ -1404,7 +1404,7 @@ int Character_GetBlockingHeight(CharacterInfo *chaa)
 
 void Character_SetBlockingHeight(CharacterInfo *chaa, int hit) {
 
-    chaa->blocking_height = hit;
+    chaa->blocking_height = static_cast<int16_t>(hit);
 }
 
 int Character_GetBlockingWidth(CharacterInfo *chaa)
@@ -1414,7 +1414,7 @@ int Character_GetBlockingWidth(CharacterInfo *chaa)
 
 void Character_SetBlockingWidth(CharacterInfo *chaa, int wid)
 {
-    chaa->blocking_width = wid;
+    chaa->blocking_width = static_cast<int16_t>(wid);
 }
 
 int Character_GetBlockingRectX(CharacterInfo *chaa)
